complete hw4
