+++ src/components/Menu/index.js
@@ -1,0 +1,39 @@
+import cn from 'classnames'
+
+import s from './style.module.css'
+import { useState } from 'react'
+const Menu = () => {
+  const [isActive, setActive] = useState(false);
+  const handleClick = () => {
+    setActive(!isActive);
+  }
+return (
+<div className={cn(s.menuContainer, {[s.active] : isActive})} onClick={handleClick}>
+  <div className={s.overlay}/>
+  <div className={s.menuItems}>
+    <ul>
+      <li>
+        <a href="#welcome">
+          HOME
+        </a>
+      </li>
+      <li>
+        <a href="#game">
+          GAME
+        </a>
+      </li>
+      <li>
+        <a href="#about">
+          ABOUT
+        </a>
+      </li>
+      <li>
+        <a href="#contact">
+          CONTACT
+        </a>
+      </li>
+    </ul>
+  </div>
+</div>
+)}
+export default Menu;
