+++ src/components/MenuHeader/index.js
@@ -1,0 +1,13 @@
+import Menu from '../Menu'
+import NavBar from '../NavBar'
+
+const MenuHeader = () => {
+  return(
+    <div>
+      <Menu />
+      <NavBar />
+    </div>
+  );
+}
+
+export default MenuHeader;
