+++ src/components/Pokemoncard/index.js
@@ -1,6 +1,7 @@
-import s from './pokemoncard.module.css'
 import { useState } from 'react';
+import cn from 'classnames';
 import cardBackSide from '../../background/card-back-side.jpeg';
+import s from './pokemoncard.module.css'
 
 export const PokemonCard = (
     {
@@ -15,24 +16,19 @@ export const PokemonCard = (
   ) =>{
   const [isActive, setActive] = useState(false);
   const handleClick = () => {
-    if (isActive){
-      setActive(false);
-    }
-    else {
-      setActive(true);
-    }
+    setActive(!isActive);
   }
 return(
   <div className={s.root} onClick={handleClick}>
-    <div className={`${s.pokemonCard} ${isActive ? s.active : ''}`}>
+    <div className={cn(s.pokemonCard, {[s.active] : isActive})}>
       <div className={s.cardFront}>
-        <div className={`${s.wrap} ${s.front}`}>
-          <div className={`${s.pokemon} ${s[type]}`}>
+        <div className={cn(s.wrap, s.front)}>
+          <div className={cn(s.pokemon, s[type])}>
             <div className={s.values}>
-              <div className={`${s.count} ${s.top}`}>{values.top}</div>
-              <div className={`${s.count} ${s.right}`}>{values.right}</div>
-              <div className={`${s.count} ${s.bottom}`}>{values.bottom}</div>
-              <div className={`${s.count} ${s.left}`}>{values.left}</div>
+              <div className={cn(s.count, s.top)}>{values.top}</div>
+              <div className={cn(s.count, s.right)}>{values.right}</div>
+              <div className={cn(s.count, s.bottom)}>{values.bottom}</div>
+              <div className={cn(s.count, s.left)}>{values.left}</div>
             </div>
             <div className={s.imgContainer}>
               <img src={img} alt={name} />
@@ -40,7 +36,7 @@ return(
             <div className={s.info}>
               <span className={s.number}>#{id}</span>
                 <h3 className={s.name}>{name}</h3>
-                <small className={s.type}>Type:
+                <small className={s[type]}>Type:
                   <span>{type}</span>
                 </small>
             </div>
@@ -48,8 +44,8 @@ return(
         </div>
       </div>
       <div className={s.cardBack}>
-        <div className={`${s.wrap} ${s.back}`}>
-          <img src={cardBackSide } alt="Сard Backed" />
+        <div className={cn(s.wrap, s.back)}>
+          <img src={ cardBackSide } alt="Сard Backed" />
         </div>
       </div>
     </div>
